Sprint3: work with FrontEnd. All paths for Ads excluding of /ads/me; /ads/{id}/image

+++ src/main/java/ga/heaven/marketplace/service/AdsService.java
@@ -15,7 +15,7 @@ public interface AdsService {
 
     int removeAds(long id);
 
-    int updateAds(int id, CreateAds createAds);
+    int updateAds(long id, CreateAds createAds);
     
     List<Ads> getAdsMe();
